requirements.txt
streamlit>=1.33
PyMuPDF>=1.24
scikit-learn>=1.4
numpy>=1.26
pandas>=2.2
fpdf2>=2.7
python-dotenv>=1.0
pytest>=8.2
